# backend_trq_bec/TRQ_BEC_API_Tres_Fivelas_v0.5.0a2/payload/docs/openapi-liberrotas-public-v0.5.0a2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TRQ-BEC LiberRotas API",
    "summary": "Fronteira autoritativa de segurança para emissão e resgate de cupons.",
    "description": "## Fronteira de segurança do LiberRotas\n\nA **TRQ-BEC LiberRotas API** é a camada autoritativa para cadastro de dispositivos,\nemissão de cupons, resgate com prova de posse e registro de evidências no ledger.\n\n### Autenticação\n\nAs operações protegidas exigem um **Firebase ID token** no esquema HTTP Bearer.\nNo botão **Authorize**, cole somente o token — a interface acrescenta `Bearer` automaticamente.\n\n### Regra criptográfica\n\nA decisão é *fail-closed*: se a verificação criptográfica não estiver íntegra,\na autorização não produz `ALLOW`. O estado do provider pode ser consultado em\n`GET /health/crypto`.\n\n### Estado desta versão\n\nA versão `0.5.0-alpha.2` mantém Ed25519 apenas como perfil de laboratório.\nML-KEM-768 e ML-DSA-65 permanecem bloqueados até a integração de um provider\nauditado e aprovado.",
    "contact": {
      "name": "Equipe TRQ-BEC / LiberRotas"
    },
    "version": "0.5.0a2"
  },
  "paths": {
...
    "/v1/trq-bec/devices/enroll": {
      "post": {
        "tags": [
          "Dispositivos"
        ],
        "summary": "Cadastrar uma chave de dispositivo",
        "description": "Vincula a chave pública do dispositivo ao sujeito autenticado pelo Firebase. O cadastro é idempotente para a mesma identidade e falha com conflito quando há tentativa de reatribuição.",
        "operationId": "enrollDevice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Dispositivo cadastrado ou já cadastrado para o mesmo sujeito.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollDeviceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Firebase ID token ausente, malformado, expirado, revogado ou inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "bearerAusente": {
                    "summary": "Cabeçalho Bearer ausente",
                    "value": {
                      "code": "AUTH_BEARER_REQUIRED",
                      "message": "AUTH_BEARER_REQUIRED"
                    }
                  },
                  "bearerInvalido": {
                    "summary": "Cabeçalho Bearer inválido",
                    "value": {
                      "code": "AUTH_BEARER_INVALID",
                      "message": "AUTH_BEARER_INVALID"
                    }
                  },
                  "tokenInvalido": {
                    "summary": "Firebase ID token inválido",
                    "value": {
                      "code": "AUTH_TOKEN_INVALID",
                      "message": "AUTH_TOKEN_INVALID"
                    }
                  },
                  "sujeitoInvalido": {
                    "summary": "Sujeito do token inválido",
                    "value": {
                      "code": "AUTH_SUBJECT_INVALID",
                      "message": "AUTH_SUBJECT_INVALID"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "A chave do dispositivo ou a chave pública já está vinculada a outra identidade.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "identificadorJaVinculado": {
                    "summary": "Identificador de chave já vinculado",
                    "value": {
                      "code": "DEVICE_KEY_BINDING_CONFLICT",
                      "message": "DEVICE_KEY_BINDING_CONFLICT"
                    }
                  },
                  "chavePublicaJaVinculada": {
                    "summary": "Chave pública já vinculada",
                    "value": {
                      "code": "DEVICE_PUBLIC_KEY_ALREADY_BOUND",
                      "message": "DEVICE_PUBLIC_KEY_ALREADY_BOUND"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "FirebaseIDToken": []
          }
        ]
      }
    },
    "/v1/trq-bec/coupons/issue": {
      "post": {
        "tags": [
          "Cupons"
        ],
        "summary": "Emitir um cupom protegido",
        "description": "Emite o envelope e o payload compacto de QR para um cupom ativo pertencente ao empreendedor. Exige claim `role=entrepreneur`, vínculo entre `issuer_id` e o Firebase UID e dispositivo previamente cadastrado.",
        "operationId": "issueCoupon",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueCouponRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Payload compacto para QR e prazo de validade do envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueCouponResponse"
                }
              }
            }
          },
          "401": {
            "description": "Firebase ID token ausente, malformado, expirado, revogado ou inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "bearerAusente": {
                    "summary": "Cabeçalho Bearer ausente",
                    "value": {
                      "code": "AUTH_BEARER_REQUIRED",
                      "message": "AUTH_BEARER_REQUIRED"
                    }
                  },
                  "bearerInvalido": {
                    "summary": "Cabeçalho Bearer inválido",
                    "value": {
                      "code": "AUTH_BEARER_INVALID",
                      "message": "AUTH_BEARER_INVALID"
                    }
                  },
                  "tokenInvalido": {
                    "summary": "Firebase ID token inválido",
                    "value": {
                      "code": "AUTH_TOKEN_INVALID",
                      "message": "AUTH_TOKEN_INVALID"
                    }
                  },
                  "sujeitoInvalido": {
                    "summary": "Sujeito do token inválido",
                    "value": {
                      "code": "AUTH_SUBJECT_INVALID",
                      "message": "AUTH_SUBJECT_INVALID"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Identidade autenticada, porém sem claim, vínculo de emissor ou dispositivo necessário.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "claimEmpreendedor": {
                    "summary": "Claim de empreendedor ausente",
                    "value": {
                      "code": "ENTREPRENEUR_CLAIM_REQUIRED",
                      "message": "ENTREPRENEUR_CLAIM_REQUIRED"
                    }
                  },
                  "vinculoEmissor": {
                    "summary": "Emissor não corresponde ao Firebase UID",
                    "value": {
                      "code": "ISSUER_BINDING_MISMATCH",
                      "message": "ISSUER_BINDING_MISMATCH"
                    }
                  },
                  "dispositivoNaoCadastrado": {
                    "summary": "Dispositivo emissor não cadastrado",
                    "value": {
                      "code": "DEVICE_NOT_ENROLLED",
                      "message": "DEVICE_NOT_ENROLLED"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cupom inexistente, inativo ou não pertencente ao emissor autenticado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "cupomIndisponivel": {
                    "summary": "Cupom não está ativo ou não pertence ao emissor",
                    "value": {
                      "code": "COUPON_NOT_ACTIVE_OR_NOT_OWNED",
                      "message": "COUPON_NOT_ACTIVE_OR_NOT_OWNED"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Provider pós-quântico indisponível ou adaptação do envelope ainda não aprovada. A emissão falha fechada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "providerNaoPronto": {
                    "summary": "Provider pós-quântico não está pronto",
                    "value": {
                      "code": "PQC_PROVIDER_NOT_READY",
                      "message": "PQC_PROVIDER_NOT_READY"
                    }
                  },
                  "adaptadorPendente": {
                    "summary": "Adaptador de envelope aguarda aprovação",
                    "value": {
                      "code": "PQC_ENVELOPE_ADAPTER_PENDING_APPROVAL",
                      "message": "PQC_ENVELOPE_ADAPTER_PENDING_APPROVAL"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "FirebaseIDToken": []
          }
        ]
      }
    },
    "/v1/trq-bec/coupons/redeem/begin": {
      "post": {
        "tags": [
          "Resgate"
        ],
        "summary": "Iniciar o resgate e emitir um desafio",
        "description": "Valida o vínculo do QR, o estado do token e o dispositivo do visitante. Cria uma operação, uma sessão e um desafio de uso único, retornando a mensagem exata que deverá ser assinada no dispositivo.",
        "operationId": "beginCouponRedemption",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeginRedemptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Operação pendente, desafio de frescor e mensagem para prova de posse.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeginRedemptionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Firebase ID token ausente, malformado, expirado, revogado ou inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "bearerAusente": {
                    "summary": "Cabeçalho Bearer ausente",
                    "value": {
                      "code": "AUTH_BEARER_REQUIRED",
                      "message": "AUTH_BEARER_REQUIRED"
                    }
                  },
                  "bearerInvalido": {
                    "summary": "Cabeçalho Bearer inválido",
                    "value": {
                      "code": "AUTH_BEARER_INVALID",
                      "message": "AUTH_BEARER_INVALID"
                    }
                  },
                  "tokenInvalido": {
                    "summary": "Firebase ID token inválido",
                    "value": {
                      "code": "AUTH_TOKEN_INVALID",
                      "message": "AUTH_TOKEN_INVALID"
                    }
                  },
                  "sujeitoInvalido": {
                    "summary": "Sujeito do token inválido",
                    "value": {
                      "code": "AUTH_SUBJECT_INVALID",
                      "message": "AUTH_SUBJECT_INVALID"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "QR inválido, vínculo inconsistente ou envelope criptográfico rejeitado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "vinculoQrInvalido": {
                    "summary": "QR não corresponde ao token armazenado",
                    "value": {
                      "code": "QR_BINDING_MISMATCH",
                      "message": "QR_BINDING_MISMATCH"
                    }
                  },
                  "tokenCriptograficoInvalido": {
                    "summary": "Envelope criptográfico inválido",
                    "value": {
                      "code": "TOKEN_CRYPTO_INVALID",
                      "message": "TOKEN_CRYPTO_INVALID"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "O dispositivo do visitante não está cadastrado para a identidade autenticada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "dispositivoNaoCadastrado": {
                    "summary": "Dispositivo de resgate não cadastrado",
                    "value": {
                      "code": "DEVICE_NOT_ENROLLED",
                      "message": "DEVICE_NOT_ENROLLED"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Referência de token desconhecida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "tokenDesconhecido": {
                    "summary": "Referência do token não encontrada",
                    "value": {
                      "code": "TOKEN_REFERENCE_UNKNOWN",
                      "message": "TOKEN_REFERENCE_UNKNOWN"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Token já utilizado, retido ou em estado incompatível com novo resgate.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "tokenNaoResgatavel": {
                    "summary": "Token não pode ser resgatado novamente",
                    "value": {
                      "code": "TOKEN_NOT_REDEEMABLE",
                      "message": "TOKEN_NOT_REDEEMABLE"
                    }
                  }
                }
              }
            }
          },
          "410": {
            "description": "Token expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "examples": {
                  "tokenExpirado": {
                    "summary": "Prazo de validade do token encerrado",
                    "value": {
                      "code": "TOKEN_EXPIRED",
                      "message": "TOKEN_EXPIRED"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "FirebaseIDToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Código de motivo legível por máquina.",
            "examples": [
              "AUTH_TOKEN_INVALID"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Mensagem de erro legível por pessoa ou operador.",
            "examples": [
              "AUTH_TOKEN_INVALID"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorResponse",
        "description": "Envelope estável de erro retornado nas fronteiras de autenticação e serviço.",
        "examples": [
          {
            "code": "AUTH_TOKEN_INVALID",
            "message": "AUTH_TOKEN_INVALID"
          }
        ]
      },
...
      "EnrollDeviceRequest": {
        "properties": {
          "device_key_id": {
            "type": "string",
            "maxLength": 160,
            "minLength": 12,
            "pattern": "^device:[A-Za-z0-9:_-]+$",
            "title": "Device Key Id",
            "description": "Identificador estável da aplicação para a chave local do dispositivo.",
            "examples": [
              "device:expo:feirante:01"
            ]
          },
          "public_key_b64u": {
            "type": "string",
            "maxLength": 43,
            "minLength": 43,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Public Key B64U",
            "description": "Chave pública Ed25519 bruta codificada em Base64URL sem preenchimento (perfil de laboratório).",
            "examples": [
              "iOZAaQkSG1nU4MV5kY_IRQAxIQ1o9S87b2m3oxygTlw"
            ]
          },
          "algorithm": {
            "type": "string",
            "const": "ED25519_LAB",
            "title": "Algorithm",
            "description": "Algoritmo de prova do dispositivo habilitado nesta versão de laboratório."
          },
          "storage_profile": {
            "type": "string",
            "const": "EXPO_SECURE_STORE_LAB",
            "title": "Storage Profile",
            "description": "Perfil esperado de proteção local da chave privada."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "device_key_id",
          "public_key_b64u",
          "algorithm",
          "storage_profile"
        ],
        "title": "EnrollDeviceRequest",
        "examples": [
          {
            "algorithm": "ED25519_LAB",
            "device_key_id": "device:expo:feirante:01",
            "public_key_b64u": "iOZAaQkSG1nU4MV5kY_IRQAxIQ1o9S87b2m3oxygTlw",
            "storage_profile": "EXPO_SECURE_STORE_LAB"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "FirebaseIDToken": {
        "type": "http",
        "description": "Firebase ID token emitido para o usuário autenticado do LiberRotas. Cole somente o valor do token; a Swagger UI acrescenta o prefixo `Bearer` automaticamente.",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token (JWT)"
      }
    }
  },
...
}